kjør rollback í db-tester via superklasse / extension
