+++ src/main/resources/db/migration/V5__mottak_status.sql
@@ -1,0 +1,1 @@
+ALTER TABLE MOTTAK ADD COLUMN GYLDIG BOOLEAN DEFAULT TRUE;
