+++ src/test/resources/db/migration/V4__mottak_status.h2
@@ -1,0 +1,1 @@
+ALTER TABLE MOTTAK ADD COLUMN GYLDIG BOOLEAN DEFAULT TRUE;
